colab_requirements.txt
grequests==0.6.0
certifi==2020.11.8
chardet==3.0.4
cryptography==3.2.1
cycler==0.10.0
Cython==0.29.21
gevent==20.9.0
greenlet==0.4.17
idna==2.10
kiwisolver==1.3.1
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.1
PyYAML==5.3.1
six==1.15.0
zope.event==4.5.0
zope.interface==5.2.0